Add more accepted caps so we can support more webcams.
Add more capabilities in the pipeline variating framerate and
video format.
Also, in order to add this new data into the pipeline in a clean
way a new private method has been added called "basicPipe" which
returns the first part of the pipeline which is common between
WebcamWidget::playFile and WebcamWdiget::recordVideo.