Force yuv/I420 in a ffmpegcolrospace after v4l2.
Seems that the negotiation between v4l2 and the rest of our pipeline
goes wrong somehow some times, so to be sure that we're getting I420
which is a known working format I've added a ffmpegcolorspace just
after v4l2src.

Also, now we're using the I420 convertion in photoGstCallback and
I had to fix the Fourcc comparation there.