Move the gallery into an togglable overlay item
This came from a brainstorm between Jens, Alex and I. It seems to make
sense, as we'll get to remove quite a bit of the clutter the main UI had
while keeping the ability to share and light-browse the resources set.
Of course, still needs work.

CCMAIL: afiestas@kde.org
CCMAIL: jens@ohyran.se