Do not reset pipeline if not needed (only when changing devices)
This part of kamoso is terribly coupled, but before we refactor it we
better get it working great.

What this commit does is changing the mode of camerabin to picture if
play is called again with the same device and only resetting the
pipeline if we change to a new device.