Use a better icon for the "Shoot" action
Summary:
The "Shoot" action right now displays a picture for an icon. This is a missed
opportunity to use a camera, which does a better job of communicating what the
action will actually do. We can also use an action from the icon theme and remove
one of the three custom icons we include, which the HIG recommends against doing
(https://hig.kde.org/style/icon.html#general-guidelines)

Test Plan: {F6530599, size=full}

Reviewers: apol, #vdg

Reviewed By: apol

Subscribers: abetts

Differential Revision: https://phabricator.kde.org/D18033