Restore timer and burst mode photo count
Summary:
The top text was always displaying the `modeInfo` text from the primary action,
which worked when the primary action jumped around. After that was fixed, this
part was never changed, so it was always displaying the `modeInfo` text for the
photo mode, which is an empty string.

This patch fixes that sonow the recording time is displayed in video mode
and the burst photo count is displayed in burst mode.

BUG: 402951
FIXED-IN 19.04.0

Test Plan: Use video mode or burst mode

Reviewers: apol

Reviewed By: apol

Differential Revision: https://phabricator.kde.org/D18200