When in Desktop view, show a desktop-style toolbar with centered actions
Summary:
One thing that has bugged me about Kamoso forever is how it insists on looking like a
mobile app when run on the desktop. This seems to be to be contrary to Kirigami's
reason for existence and does not showcase the toolkit's excellent convergence
features.

This patch resolves that by using the default auto-switching feature and therefore
displays a desktop-style toolbar when run on the desktop. To improve the appearance, it
makes use of a new optional feature in Kirigami added in D23845 to center the actions
when in desktop toolbar mode.

BUG: 402948
FIXED-IN: 19.12.0

Depends on D23845

Test Plan:
Apply D23845 to Kirigami, then apply this to Kamoso and run the app.

Before: {F7338519}

After: {F7338520}

When D23845 is not applied, the actions are right-aligned, but nothing bad happens, so I opted not to increase the required Frameworks version in the CMakeLists.txt file.

Reviewers: apol, #vdg, ndavis

Reviewed By: #vdg, ndavis

Differential Revision: https://phabricator.kde.org/D23847