Port away from udi+path to objectid
At the moment we're just using the udi to identify the device and we're
not even using v4l2src directly, since gst is creating the source on its
own and relying on pipewire.

This simplifies the infrastructure by removing an unused attribute (the
path) and using the right identifier.